ログアウト機能

+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\TaskController;
 use App\Http\Controllers\UserController;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -21,16 +22,17 @@ Route::get('/', function () {
 });
 
 
-// Route::resource("/tasks", TaskController::class);
-Route::get("/tasks/index", [TaskController::class, "index"])->name("tasks.index");
-Route::get("/tasks/create", [TaskController::class, "create"])->name("tasks.create");
-Route::post("/tasks/store", [TaskController::class, "store"])->name("tasks.store");
-Route::get("/tasks/{task}", [TaskController::class, "edit"])->name("tasks.edit");
-Route::put("/tasks/{task}", [TaskController::class, "update"])->name("tasks.update");
-Route::delete("/tasks/{task}", [TaskController::class, "destroy"])->name("tasks.destroy");
+// Route::resource("/tasks", TaskController::class)->except("show");
+Route::get("/tasks/index", [TaskController::class, "index"])->name("tasks.index")->middleware("auth");
+Route::get("/tasks/create", [TaskController::class, "create"])->name("tasks.create")->middleware("auth");
+Route::post("/tasks/store", [TaskController::class, "store"])->name("tasks.store")->middleware("auth");
+Route::get("/tasks/{task}", [TaskController::class, "edit"])->name("tasks.edit")->middleware("auth");
+Route::put("/tasks/{task}", [TaskController::class, "update"])->name("tasks.update")->middleware("auth");
+Route::delete("/tasks/{task}", [TaskController::class, "destroy"])->name("tasks.destroy")->middleware("auth");
 
 Route::get("/users/create", [UserController::class, "create"])->name("users.create");
 Route::post("/users", [UserController::class, "store"])->name("users.store");
 
 Route::get("/users/login", [AuthController::class, "loginForm"])->name("users.login");
 Route::post("/users/login", [AuthController::class, "login"]);
+Route::post("/users/logout", [AuthController::class, "logout"])->name("users.logout");
